ajout component SearchGameDisplay et mise en palce sur filtre gauche dessus
