+++ frontend/src/context/MyContext.jsx
@@ -13,6 +13,7 @@ export function ContextProvider({ children }) {
   const [fetchAllCategories, setFetchAllCategories] = useState([]);
   const [fetchAllGames, setFetchAllGames] = useState([]);
   const [fetchCategoryByGame, setFetchCategoryByGame] = useState([]);
+  const [searchResult, setSearchResult] = useState([]);
   const [editor, setEditor] = useState("");
   const [language, setLanguage] = useState("");
   const [player, setPlayer] = useState("");
@@ -44,6 +45,11 @@ export function ContextProvider({ children }) {
     axios.get(url).then((response) => setFetchCategoryByGame(response.data));
   };
 
+  const searchList = () => {
+    const url = "http://localhost:8000/api/game";
+    axios.get(url).then((response) => setSearchResult(response.data));
+  };
+
   return (
     <Provider
       value={{
@@ -69,6 +75,8 @@ export function ContextProvider({ children }) {
         setDuration,
         getCategoryByGame,
         fetchCategoryByGame,
+        searchList,
+        searchResult,
       }}
     >
       {children}
