--- frontend/src/css/GameSearched.css
@@ -1,16 +1,0 @@
-.main_gameSearched {
-  height: auto;
-  display: flex;
-  justify-content: flex-start;
-  background-color: white;
-  border: 1px solid black;
-}
-
-.gameSearched_display {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 5vh;
-  margin-right: 1vw;
-  margin-left: 1vw;
-  gap: 1vw;
-}
